update node version to LTS
