+++ Dockerfile
@@ -1,4 +1,4 @@
-FROM node:8.12.0
+FROM node:10.13.0
 
 RUN npm install --production
 RUN npm start
