+++ package.json
@@ -1,11 +1,11 @@
 {
-  "name": "starterpack",
+  "name": "strvproject",
   "version": "0.0.1",
   "author": "diegobarrera",
-  "descripction": "Template project for a Node.js API",
+  "descripction": "STRV address book project",
   "license": "MIT",
   "engines": {
-    "node": "^8.12.0",
+    "node": "^10.13.0",
     "npm": "^6.4.0"
   },
   "scripts": {
@@ -18,7 +18,6 @@
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
     "express": "~4.16.0",
-    "json-api-serializer": "^1.13.0",
     "mocha": "^5.2.0",
     "morgan": "^1.9.1",
     "winston": "^3.1.0"
@@ -30,6 +29,7 @@
     "eslint-plugin-node": "^7.0.1",
     "eslint-plugin-promise": "^4.0.1",
     "eslint-plugin-standard": "^4.0.0",
+    "nodemon": "^1.18.7",
     "standard": "^12.0.1"
   }
 }
