+++ src/app.js
@@ -3,6 +3,7 @@ const express = require('express')
 const cookieParser = require('cookie-parser')
 const morgan = require('morgan')
 const winston = require('../src/configuration/winston')
+const { errorHandler } = require('../src/middlewares/errorHandler')
 const app = express()
 
 const indexRouter = require('../src/routes/index')
@@ -18,18 +19,14 @@ app.use(cookieParser())
 
 app.use('/api', indexRouter)
 
-const errorHandler = (err, req, res, next) => {
-  // set locals, only providing error in development
-  res.locals.message = err.message
-  res.locals.error = req.app.get('env') === 'development' ? err : {}
-
-  // winston error logging
-  winston.error(`${err.status || 500} - ${err.message} - ${req.originalUrl} - ${req.method} - ${req.ip}`)
-
-  // render the error page
-  res.status(err.status || 500)
-  res.send('error')
-}
 app.use(errorHandler)
+
+process.on('unhandledRejection', (reason, p) => {
+  throw reason
+})
+process.on('uncaughtException', (error) => {
+  errorHandler(error)
+  process.exit(1)
+})
 
 module.exports = app
