+++ src/middlewares/errorHandler.js
@@ -1,0 +1,18 @@
+'use strict'
+
+const winston = require('../configuration/winston')
+
+const errorHandler = (err, req, res, next) => {
+  res.locals.message = err.message
+  res.locals.error = req.app.get('env') === 'development' ? err : {}
+
+  winston.error(`${err.status || 500} - ${err.message} - ${req.originalUrl} - ${req.method} - ${req.ip}`)
+
+  res.status(err.status || 500)
+  res.json({
+    error: err.message
+  })
+}
+module.exports = {
+  errorHandler
+}
